+++ odoo/local-src/specific_zetes/models/__init__.py
@@ -2,4 +2,4 @@ from . import res_users
 from . import stock_picking
 from . import res_partner
 from . import zetes_logger
-from . import report_stock_quant_bylocation
+from . import stock_picking_type