+++ odoo/local-src/specific_zetes/tests/test_assignment.py
@@ -22,7 +22,7 @@ class TestAssignemnt(ZetesTest):
             'is_passport_required': True,
         })
 
-        self.assertEqual(self.picking.zetes_picking_type,
+        self.assertEqual(self.picking.picking_type_id.zetes_picking_type,
                          constants.PICKING_ASSIGNMENT)
 
         # Search for a picking